#1272 Moved test into goloop.
